refactor: seperate controllers and routes for articles and products
