+++ http/products.http
@@ -5,16 +5,13 @@ GET http://localhost:3000/products
 GET http://localhost:3000/products/1e112c02-7c92-4d52-b76e-485e41e3f64d
 
 ###
-GET http://localhost:3000/products/1e112c02-7c92-4d52-b76e-485e41e3f64d/comments
-
-###
 GET http://localhost:3000/products?orderBy=oldest
 
 ###
 GET http://localhost:3000/products?orderBy=recent&page=1&pageSize=2
 
 ###
-GET http://localhost:3000/products?keyword=샥
+GET http://localhost:3000/products?keyword=고급스러운
 
 ###
 POST http://localhost:3000/products
@@ -26,11 +23,11 @@ Content-Type: application/json
   "category": "ELECTRONICS",
   "price": 100000,
   "stock": 10,
-  "tags": ["태그는 5글자 이내여야 함", "신상", "할인"]
+  "tags": ["태그", "신상", "할인"]
 }
 
 ###
-PATCH  http://localhost:3000/products/5f449bfe-4743-44e5-961e-6d0ea08b5e4e
+PATCH  http://localhost:3000/products/997b2f43-01d1-4ed8-a094-39cd4a4406c8
 Content-Type: application/json
 
 {
